der: add `indefinite` flag to `Length` (#1894)
Tracks whether decoded lengths were `indefinite`, which is needed to
decide how to decode the field's value.

Also adds a check that the constructed bit is set when an indefinite
length is encountered in a `Header`.

+++ pkcs5/src/pbes2/kdf/salt.rs
@@ -33,7 +33,7 @@ impl Salt {
 
         Ok(Self {
             inner,
-            length: Length::new(slice.len() as u16),
+            length: Length::new(slice.len() as u32),
         })
     }
